wordclouds produced from IMDB's quotes data
these word clouds are created for the decades spanning through the 50s
until present day. for each decade, the frequency of each word from a
list of words consisting of basic Spanish vocabulary, words of
prejudice geared toward Latinx, and popular markers of different Latinx
cultures is measured. it’s not a question of whether Latinx culture is
present but rather a question of what aspects are used to construct the
on-screen Latinx identity